requirements.txt
# Python version: 3.12.5
# Pip version: 24.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altgraph==0.17.4
attrs==24.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
browser-cookie3==0.19.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
curl_cffi==0.7.1
et-xmlfile==1.1.0
frozenlist==1.4.1
g4f==0.3.2.4
idna==3.8
lxml==5.3.0
lz4==4.3.3
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
openpyxl==3.1.5
packaging==24.1
pefile==2024.8.26
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyPDF2==3.0.1
python-docx==1.1.2
pywin32-ctypes==0.2.3
requests==2.32.3
rich==13.8.0
setuptools==73.0.1
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4